Swap column order, rename utils to server
